pyproject.toml
[tool.poetry]
name = "aia-mcp"
version = "0.2.0-alpha.1"
description = "Servidor MCP para el agente aia (amanda-IA)"
readme = "README.md"
packages = [
    {include = "temperatura"},
    {include = "wahapedia"},
    {include = "tinaja"},
    {include = "mcp_cli"},
]

[tool.poetry.scripts]
mcp = "mcp_cli.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
mcp = ">=1.0"
httpx = ">=0.27"
beautifulsoup4 = ">=4.12"
paho-mqtt = ">=2.0"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
